+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.2 (unreleased)
 
 - Add `sdwan_application_aware_routing_policy_definition` resource and data source
+- Add `sdwan_cflowd_policy_definition` resource and data source
 
 ## 0.2.1